Added support for setting register values.
This is still quite experimental. To be more specific, I think kvm_set_vcpureg
is likely to be broken. The reason for this is that the kernel only supports
setting all the registers at once and for us to be able to set only a single
register we need to retrieve the old values first. Of course, both of these
operations need to be done without the vcpu continuing execution while the
operations are happening. Maybe we should explicitly pause the vcpu before the
operations.